// src/types/mod.ts
export interface Mod {
  slug: string;
  originalFilename: string;
  title: string;
  image: string;
  duration: string;
  category: string;
  status: string;
  developer: string;
  translator: string;
  revisor: string;
  mobile: string;
  mobileCredits: string;
  description: string[];
  pcDownload: string;
  apkDownload: string;
}